sync: create dedicated smart sync func
The smart sync logic takes up about 45% of the overall Execute func
and is about 100 lines of code.  The only effect it has on the rest
of the code is to set the manifest_name variable.  Since this func
is already quite huge, split out the smart sync logic out.

Change-Id: Id861849b0011ab47387d74e92c2ac15afcc938ba