repo: use encoding parameter for subprocess.Popen
During testing of the initial repo bootstrap file using Python3,
2 errors were encountered relating to Python3 API changes for
subprocess.Popen.  If an encoding parameter is not specified
the returned value is a byte-like object and not a str.

The first error was during the initial setup of MAINTAINER_KEYS:
File "<snip>/bin/repo", line 508, in SetupGnuPG
    proc.stdin.write(MAINTAINER_KEYS)
TypeError: a bytes-like object is required, not 'str'

The second error happened after fixing the first error:
  File "<snip>/bin/repo", line 691, in _Verify
    m = re.compile(r'^(.*)-[0-9]{1,}-g[0-9a-f]{1,}$').match(cur)
TypeError: cannot use a string pattern on a bytes-like object

NOTE: Adding the encoding parameter is an API change which is
incompatible with Python2.

Signed-off-by: Michael Scott <mike@foundries.io>
Change-Id: I1863d6e21ba9546290c380bb5d405fc87e252733