include: Put 'remote' and 'default' elements to namespaces
The repo client could not support top manifest file include
low level manifest files with 'remote' and 'default'
elements which has the same name but different attributes.

And the repo client throw errors if one manifest file is
included more than once through different manifest files.

From the "repo Manifest Format" the 'include' element is
described as:

    Normal rules apply for the target manifest to include
    - it must be a usable manifest on its own

which says that the target manifest to include must be a
usable manifest on its own, it seems like each manifest
file should has it's own namespace of it's 'remote' and
'default' elements, and what we found is repo client does
not acts like that.

Each element of manifest file should has it's namespace,
so the repo client will throw errors only if the manifest
file is not a usable file on its own. And the repo client
should not complains if one manifest file is included more
than once.

Signed-off-by: sunchao <1987.sunchao@gmail.com>
Change-Id: Ieca98ad96c63df7c19c41f593fa75147bfe08d5b