sync: add option to keep local workspace manifest
The use case is any situation where your manifest does
not exist on server, but where you still want to do
full sync for the projects, without having your
workspace manifest switched to other branch or
forwarded to latest or similar.
Since not all organizations use frequent snap generation,
a sync to a historical manifest should still be possible
for them, or to have a synced workspace when integrating
something together that has not been pushed upstream yet.
Changes can exist locally on a manifest that is behind
head, meaning not requiring rebase to latest.

Tested using:
  $ cd .repo/manifests/
  $ git checkout <any hash 1>
  $ <do local modifications>
  $ repo sync -w
  $ git checkout <any hash 2>
  $ repo sync -w

Change-Id: I0c9773aa8bc5876813a2e7d7fec697abcb2d9e94