Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> i don't think this is a common enough operation to warrant a short flag
> 
> for the long flag, "--workspace-mp" doesn't seem all that clear to me for what it's actually doing.  how about --no-manifest-update or something like that ?

Thanks for the feedback, Mike!

I went for "which is the short letter least likely to be needed for some other short option", came up with "w", then tried to construct a meaningful long name, but most got so long that it pushed the columns in the Usage print, and I ended up with the not so helpful "workspace-mp", sorry about that. =)

If I switch to "keep" as first word maybe, and remove short option, then at least maybe in the future, the short "k" is not already taken like "n" is?

Patch-set: 1