Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> i'm not keen on trying to support anything other than UTF-8 by design.  is it really necessary to add that ?  if people are trying to use non-UTF-8 locales, they need to fix their env imo.

Windows 10 doesn't fully support UTF-8 locales (old builds of Windows 10 don't support them at all) and by default this is completely disabled because it's experimental:
https://superuser.com/questions/1033088/is-it-possible-to-set-locale-of-a-windows-application-to-utf-8
For example my current locale is cp1252, in Windows 10.0.18363.592.

On the other hand the platform_utils.putenv and the platform_utils.getenv are not really needed and the environment dictionary (the env variable) could always be updated directly. We just need to make sure we use only str strings. This will work properly in both Python 2 and 3.

Eventually the core of the repo fix for Windows is removing the _setenv() function (that in Python 3 always encodes string into bytes) and working with the env directly. This is what I did in my CL:
https://gerrit-review.googlesource.com/c/git-repo/+/251832/1/git_command.py

The _setenv() function and other calls of encode() of the put into the env values were added by this commit:
https://gerrit-review.googlesource.com/c/git-repo/+/19219
Interesting that 5 years after that it was partially disabled (in the repo launcher) but nobody seems to suffer from that in Windows:
https://gerrit-review.googlesource.com/c/git-repo/+/80225

It made a strange code:

  try:
    env['GNUPGHOME'] = gpg_dir
  except UnicodeEncodeError:
    env['GNUPGHOME'] = gpg_dir.encode()

Can the UnicodeEncodeError ever happen here? Maybe in some really old Python that already isn't supported?

Patch-set: 3
Reviewer: Gerrit User 1094845 <1094845@173816e5-2b9a-37c3-8a2e-48639d4f1153>