Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > We just need to make sure we use only str strings. This will work properly in both Python 2 and 3.
> 
> unfortunately, it isn't that simple.  using strings means you have an encoding.  please see the file comments in platform_utils.putenv which outlines these issues.
> 
> the only real time you can bypass the putenv help and access the environ directly is when you have a string with only ASCII content.  it'll convert fine with UTF-8 & cp1252/ISO-8859-1.
> 
> > but nobody seems to suffer from that in Windows
> 
> that is incorrect.  the whole reason i started this CL was because i had reports that the existing env encoding hacks were failing under Windows w/Python 3.
> 
> i think it's more that the Windows userbase of repo is quite small, and the software tends to be frozen/ancient.  no one touches things until it blows up.

Dear 
There are more Windows 3 based Python 3 user bases than you think. And best of all, Ubuntu will be upgrading to Python 3 in the near future, so I think we should try to make improvements regardless of OS platform. We are looking forward to seeing a lot of developers in our office.
Based on your comments, we hope that the fix will be corrected by additional patchsets and that the underlying issues will be resolved.

Patch-set: 3
CC: Gerrit User 1086523 <1086523@173816e5-2b9a-37c3-8a2e-48639d4f1153>