Fixing forall subcommand for Py3
Execution of 'repo forall -p -c' doesn't work with Py3 and ends up
with an error:

Got an error, terminating the pool: TypeError: can only concatenate
str (not "bytes") to str

That's fixed by the first change in the patch. After that's fixed,
another error appears:

Got an error, terminating the pool: TypeError: write() argument must
be str, not bytes

That's fixed by the second change in the patch. Changes introduced by
this patch work well with Py3 as well as with Py2.

Change-Id: Ice01aaa1822dde8d957b5bf096021dd5a2b7dd51