Fix `repo init -c`
The init subcmd understood '-c', but the repo wrapper did not, so you
could only use it on existing trees. Since the purpose of '-c' is so
that you don't have to download the manifest branches that you don't
care about, it's not that useful if you always have to download them
first.

This is a fairly large diff for what's actually changing, since we need
to delay the creation of the current-branch option until we know whether
we're in 'init' or 'gitc-init'.

Test: repo help init
Test: repo help gitc-init
Test: repo init -u https://android.googlesource.com/platform/manifest -c
Test: repo init -u https://android.googlesource.com/platform/manifest --current-branch
Change-Id: I259c0a27a831aee743084dfe621e93add5ad3eed