Update patch set 2
Patch Set 2:

in order to make sure it's atomic, it'd have to create a tempfile in the same dir as cookiefile, and then do a rename.  i think that'd work even on networked ~, but i guess depends on the FS.

even then, it'll be a bit noisy when syncing hundreds of repos ... constantly thrashing on the file to do nothing is silly.

seems like curl should be intelligent here (don't write updates if the cookiejar doesn't change, and if it does, do it atomically).  but we can't wait for that to propagate out.  i'll see if filing a bug with them upstream helps at least.

Patch-set: 2