project: set core.repositoryFormatVersion=1 when using extensions
When using extensions, make sure we set the git repo format version
so git knows to check the extension compatibility.  If git ever adds
version format 2, we'll want to add a proper helper here so different
paths don't clobber each other, but it's not clear if/when that will
ever happen.  So be lazy for now.

Change-Id: I9ab6c32d92fe2b8e5df6e2b080ca71556332e909