Expose upstream and dest-branch attributes through environment
Recent changes to ensure we're actually reading from snapshot
manifests properly have exposed several bugs in our assumptions
about manifest files.  Mainly that the revision field for a project
does _not_ have to refer to a ref, it can just be a commit hash.

Several places assume that the revision field can be parsed as a
ref to get the branch the project is on, which isn't true.  To fix
this we need to be able to look at the upstream and dest-branch
attributes of the repo, so we expose them through the environment
variables set in `repo forall`.

TESTS=manual 'repo forall' run

Change-Id: I2c039e0f4b2e0f430602932e91b782edb6f9b1ed