Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> it is difficult to use unsigned commits by default, but that's kind of WAI.  people shouldn't be using unverified code on their systems.

We don't have the infrastructure set up to make sure that all our code is GPG signed, though

> if you want to pin to an older repo release, you can do that now w/out disabling verification.

That didn't work so good for "stable", but in this particular case it was mostly about features
 
> if you want to fork repo and make your own changes, then you can set up your own signed tags to support that.

You mean fork the launcher, to add our own keys ? That's a possibility, was hoping for config


Hopefully I will be able to clean up the code enough to contribute it here instead, will see.

But I take it that adding a environment variable for this setting is out, so will Abandon.

Patch-set: 2