Update patch set 1
Patch Set 1:

Hi Mike,
This is my first contribution so I hope I've done it right...
It is a minor change and I have tested it with the latest sources.
If you think it is fine, is it also possible this change to go on
top of v1.13.11(I am not sure which exactly branch is used)?
Thanks!

Patch-set: 1
Reviewer: Gerrit User 1007145 <1007145@173816e5-2b9a-37c3-8a2e-48639d4f1153>