Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> we don't implement hooks directly like this.  plus, `repo` is project agnostic, so it can never assume that it's in an Android manifest (which is what this seems to be doing).
> 
> please see the documentation:
> https://gerrit.googlesource.com/git-repo/+/HEAD/docs/repo-hooks.md
> 
> i'll point out that the repohooks project already has a clang-format that triggers are `repo upload` time:
> https://android.googlesource.com/platform/tools/repohooks/#builtin-hooks
> 
> while there is a request for adding a pre-commit stage in general:
> https://crbug.com/gerrit/12378
> 
> i don't think people would ever want clang-format running at the `git commit` stage.  that's way too restrictive and would significantly slow things down.

Hi Mike,

Thank you for looking into this. We have precommit auto formatting installed on several internal projects (e.g. ag/10925459) and are looking for a solution to find a centralized place to store this hook so people have the option to install it. Do you have any suggestions on how we can go about doing that? Can we have this as option just like "cpplint=true" in PREUPLOAD.cfg? Some thing like "precommit_format=true"?

Also this script only performs clang-format on DIFFs, not the entire project, so it's not slow at all. 

We are aware clang_format can be done on repo upload but we believe for some, if not a lot, users they prefer to have it auto formatted on precommit, so repo upload's auto formatting don't catch them off guard.

Patch-set: 1