Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> please see the bug i referenced above.  that is the only avenue for centralizing precommit hook support in repo itself.  we def will not add any code that assumes Android checkouts to repo, and most likely won't add tool-specific logic either.  we have the repohooks mechanism specifically to offload these kinds of policy decisions.

I see. We put """
[Hook Scripts]
check_pre_commit_hook_installed = sh tools/pre-upload-check
""" in ag/10925459. Basically the script checks if user has our pre-commit installed and ask them if they want to install it. Is it possible we store this in repohooks project as have it as a built-in hook for everyone who has repo hooks?

Patch-set: 1