Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> sorry, but i don't understand what you're asking
> 
> this project (git-repo) is not Android specific.  it has a generic hook mechanism for `repo upload` time, but it is up to each project to implement that hook/framework.  Android uses such a project, but that is maintained outside of here.

Sorry for the confusion. 

To clarify, I am asking if we could add a built-in hook under repo hooks (tools/repohooks project) that triggers at "repo upload" time. Of course, users will have the option to enable/disable it, just like they to with cpplint=true/false builtin hook. And what that hooks does is that it installs a precommit hook in their projects' .git/hooks folder. We are not indenting to do this in git-repo project.

Patch-set: 1