Fix how we format the full destination branch when uploading.
If the dest-branch attribute is set in the project manifest, then
we need to push to that branch.  Previously, we would unconditionally
pre-pend the refs/heads prefix to it.  The dest-branch attribute is
allowed to be a ref expression though, so it may already have it.

Simple fix is to check if it already has the prefix before adding it.

BUG: crbug.com/gerrit/12770
TEST: manual
Change-Id: I45d6107ed6cf305cf223023b0ddad4278f7f4146