launcher: set script shebang to python3
For Debian Bullseye (to be released soon in early 2022) and Ubuntu Focal
(20.04), except installing python-is-python2 [1] or
python-is-python3 [2], there won't be a symbolic link for
/usr/bin/python.

Besides, launcher is the last script wihout python3 shebang after latest
two patches after v2.10 releasing. So hoping we can migrate to python3
completely after this one.

[1] https://packages.debian.org/bullseye/python-is-python2
[2] https://packages.debian.org/bullseye/python-is-python3

Signed-off-by: Roger Shimizu <rosh@debian.org>
Change-Id: I896be67cf70d76c765f46d3f88e4005989ba2a89